fix: move results mock utils to root folder
