+++ data/results/results-impl/src/test/kotlin/it/battagliandrea/formula1/data/results/impl/repository/ResultsRepositoryTest.kt
@@ -10,8 +10,8 @@ import it.battagliandrea.formula1.core.network.api.models.BaseResponse
 import it.battagliandrea.formula1.core.resource.Resource
 import it.battagliandrea.formula1.core.test.MainDispatcherRule
 import it.battagliandrea.formula1.data.results.api.repository.IResultsRepository
+import it.battagliandrea.formula1.data.results.impl.MockUtils
 import it.battagliandrea.formula1.data.results.impl.datasource.ErgastApiContract
-import it.battagliandrea.formula1.data.results.impl.models.MockUtils
 import kotlinx.coroutines.test.runTest
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
